Deterministic-inference fixes for hybrid-GDN (mamba/GDN) models
Ports from the TML monorepo det-inference work:
- fla/chunk_delta_h: promote state-pool indexing to int64; index * stride
  overflows int32 past slot 4096 (illegal memory access / cross-slot
  state corruption, reproducible with a >4096-slot pool).
- fla/chunk_o: pin the output tile under batch-invariant mode so tiny
  extends match the same tokens inside a large batch.
- scheduler/schedule_policy: align chunked-prefill truncation to the
  mamba cache chunk size (LCM'd with the attention split size) so hybrid
  SSM prefill numerics don't depend on budget-driven chunking.
- new --disable-finished-mamba-cache, auto-set under
  --enable-deterministic-inference: keep decode-program mamba tips out
  of the radix tree; prefix hits on them diverge from a prefill
  recompute. Guarded by a Qwen3-Next radix-tips regression test.

+++ python/sglang/srt/hardware_backend/xpu/kernels/fla/chunk_delta_h.py
@@ -111,7 +111,9 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64_k_loop(
             block_shape=(BT, BV),
         )
 
-    index = tl.load(initial_state_indices + i_n).to(tl.int32)
+    # int64: index * stride_h overflows int32 on large state pools and the
+    # wrap silently reads/writes another slot's state
+    index = tl.load(initial_state_indices + i_n).to(tl.int64)
     h0 = initial_state + index * stride_h
     ht = initial_state + index * stride_h
     if USE_INITIAL_STATE:
